// eTrade/src/app/models/categories.ts
// export interface CategoryResponse {
//   Kategoriler: Category[];
// }
// export interface Category {
//   status: boolean;
//   message: string;
//   categories: Categories[];
// }
export interface Categories {
  categoryId: number;
  categoryName: string;
  categoryDetail: string;
  categoryImage: string;
  categoryIcon :string;
}